128 bit field support

+++ src/crypto/mod.rs
@@ -3,4 +3,4 @@ pub mod hash;
 mod merkle;
 pub use merkle::{ MerkleTree, BatchMerkleProof };
 
-pub type HashFunction = fn(&[u64], &mut [u64]);
+pub type HashFunction = fn(&[u8], &mut [u8]);
